Remove cache loaders
