+++ cache/src/main/kotlin/world/gregs/voidps/cache/CacheLoader.kt
@@ -1,6 +1,5 @@
-package world.gregs.voidps.cache.memory.load
+package world.gregs.voidps.cache
 
-import world.gregs.voidps.cache.Cache
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.RandomAccessFile
@@ -8,17 +7,17 @@ import java.io.RandomAccessFile
 interface CacheLoader {
 
     fun load(path: String, xteas: Map<Int, IntArray>? = null, threadUsage: Double = 1.0): Cache {
-        val mainFile = File(path, "${FileCacheLoader.CACHE_FILE_NAME}.dat2")
+        val mainFile = File(path, "${FileCache.CACHE_FILE_NAME}.dat2")
         if (!mainFile.exists()) {
             throw FileNotFoundException("Main file not found at '${mainFile.absolutePath}'.")
         }
         val main = RandomAccessFile(mainFile, "r")
-        val index255File = File(path, "${FileCacheLoader.CACHE_FILE_NAME}.idx255")
+        val index255File = File(path, "${FileCache.CACHE_FILE_NAME}.idx255")
         if (!index255File.exists()) {
             throw FileNotFoundException("Checksum file not found at '${index255File.absolutePath}'.")
         }
         val index255 = RandomAccessFile(index255File, "r")
-        val indexCount = index255.length().toInt() / Archive.INDEX_SIZE
+        val indexCount = index255.length().toInt() / ReadOnlyCache.INDEX_SIZE
         return load(path, mainFile, main, index255File, index255, indexCount, xteas, threadUsage)
     }
